fix littlefs

+++ App/Recorder/LFS/lfs_adapter.hpp
@@ -5,4 +5,4 @@
 extern lfs_t lfs;
 extern lfs_file_t file;
 
-int initLittleFs(void);
+int8_t initLittleFs(void);
